+++ src/loaders.py
@@ -3,7 +3,7 @@
 import json
 from pathlib import Path
 
-from src.book_names import get_book_id
+from src.books import get_book_id
 from src.db import database
 from src.models import Translation, Verse
 from src.text_utils import normalize, remove_diacritics
